Lower the min value of tcp.delay_ack_msec
Lower the min value of "tcp.delay_ack_msec" from 20ms to 1ms to be able
to decrease the time before acks are sent.

The min value was set to 20ms in the because lower values caused
performance issues. The tcp timer handling was changed during the
latest TCP/IP stack uplift. The tcp timer tick resolution was fixed to
1ms and delay_ack_msec does not change the resolution anymore. There is
no additional tcp timer handling with the change of delay_ack_msec value.
So there should be no performance issues with allowing delay_ack_msec to
be set to 1ms.