Align ASTF IP address and flow count split
For ASTF templates, the distribution of IP addresses and flow counts
were previously independent. In cases where the numbers of IPs were not
a multiple of the number of threads, it sometimes led to that not all
IPs were used. The reason for this was that residue IPs where always put
on the same threads, but the residue flow count was randomly
distributed. The result was that a thread with fewer IPs could get a
higher flow count, and a thread with more IPs could get a lower flow
count. The first IPs on the thread with fewer IPs were then repeated,
and the last IPs on the thread with more IPs were never used.

With this change the IPs and the flow counts are distributed in the same
way based on the random distribution previously used for flow counts. In
cases where the number of IPs is equal to the flow count, all IPs will
now be used.